back/requirements.txt
fastapi[standard]
transformers
torch
torchvision
torchxrayvision
pillow
flask
tensorflow
sqlalchemy
psycopg2-binary
alembic
